refactor: post와 review 도메인 분리
