+++ backend/src/main/java/com/pickgo/domain/post/review/entity/Review.java
@@ -1,6 +1,7 @@
-package com.pickgo.domain.post.entity;
+package com.pickgo.domain.post.review.entity;
 
 import com.pickgo.domain.member.entity.Member;
+import com.pickgo.domain.post.post.entity.Post;
 import com.pickgo.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
