+++ backend/src/main/java/com/pickgo/domain/post/review/repository/PostReviewRepository.java
@@ -1,6 +1,6 @@
-package com.pickgo.domain.post.repository;
+package com.pickgo.domain.post.review.repository;
 
-import com.pickgo.domain.post.entity.Review;
+import com.pickgo.domain.post.review.entity.Review;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
